requirements.txt
Flask
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
Flask-admin
email-validator==1.1.3
emailvalidator==0.3
mysql-connector-python
pymysql
cryptography
WTForms
Pillow
secrets